refactor: Enable debug mode and update authorization URL in auth.js
This commit enables the debug mode in the auth.js file by setting the `debug` option to `true`. It also updates the authorization URL for the `INSA` OAuth provider to `https://cas.insa-cvl.fr/cas/oidc/oidcAuthorize` and adds the `scope` parameter with the value `openid profile`. These changes improve the functionality and debugging capabilities of the authentication system.